business-assistant-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyodbc==4.0.39
python-dotenv==1.0.0
aiohttp==3.9.1
python-telegram-bot==20.7
pydantic==2.5.0
python-multipart==0.0.6